mcp/requirements_mcp.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
ansible==11.0.0
ansible-core==2.18.0
anyio==4.8.0
aodhclient==3.6.0
appdirs==1.4.4
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
attrs==24.2.0
Authlib==1.6.1
autopage==0.5.2
babel==2.16.0
backoff==2.2.1
bcrypt==4.3.0
Beaker==1.12.1
beautifulsoup4==4.13.4
blinker==1.9.0
blivet==3.11.0
blivet-gui==2.6.0
boto3==1.39.11
botocore==1.39.11
Brlapi==0.8.5
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
chromadb==1.0.15
click==8.1.7
cliff==4.7.0
cmd2==2.5.4
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.11.2
crewai-tools==0.0.1
crypt_r==3.13.1
cryptography==43.0.0
cssselect==1.1.0
cupshelpers==1.0
cyclopts==3.22.5
dasbus==1.7
dataclasses-json==0.6.7
dbus-python==1.3.2
debtcollector==3.0.0
decorator==5.1.1
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnf==4.23.0
dnspython==2.7.0
docker==7.1.0
docstring-parser==0.15
docutils==0.22
dogpile.cache==1.3.3
durationpy==0.10
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fasteners==0.19
fastmcp==2.11.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.18.0
flatbuffers==25.2.10
fros==1.1
frozenlist==1.7.0
fsspec==2025.7.0
futurist==3.0.0
git-review==2.4.0
gnocchiclient==7.1.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
humanize==4.9.0
idna==3.7
importlib_metadata==8.7.0
importlib_resources==6.5.2
inkex==1.4.0
instructor==0.5.2
ipython==9.4.0
ipython_pygments_lexers==1.1.1
iso639==0.1.4
iso8601==2.1.0
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
json5==0.12.0
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
keyring==25.6.0
keystoneauth1==5.8.0
kubernetes==33.1.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.0.5
langchain-text-splitters==0.0.2
langsmith==0.1.147
langtable==0.0.69
lazy-object-proxy==1.11.0
legacy-cgi==2.6.1
libcomps==0.1.21
libdnf==0.74.0
libvirt-python==10.6.0
litellm==1.74.3
loguru==0.7.3
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.12.3
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.6.3
murano-pkg-check==0.3.0
mypy_extensions==1.1.0
netaddr==1.3.0
netifaces==0.11.0
networkx==3.4.2
nftables==0.1
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.22.0
openai==1.98.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openpyxl==3.1.5
openstackclient==4.0.0
openstacksdk==4.1.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==3.1.0
oslo.concurrency==6.1.0
oslo.config==9.6.0
oslo.context==5.6.0
oslo.i18n==6.4.0
oslo.log==6.1.2
oslo.serialization==5.5.0
oslo.utils==7.4.0
osprofiler==4.2.0
overrides==7.7.0
packaging==23.2
parse==1.20.2
parso==0.8.4
Paste==3.10.1
pathable==0.4.4
pbr==6.1.0
pdfminer.six==20250506
pdfplumber==0.11.7
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
pipx==1.7.1
platformdirs==4.3.7
pluggy==1.5.0
ply==3.11
portalocker==2.7.0
posthog==5.4.0
prettytable==0.7.2
productmd==1.45
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==3.0.2
pyenchant==3.2.2
Pygments==2.18.0
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.10.1
pykickstart==3.58
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyparted==3.13.0
pypdfium2==4.30.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject-api==1.9.0
pyproject_hooks==1.2.0
PyQt6==6.8.1
PyQt6-WebEngine==6.8.0
PyQt6_sip==13.9.1
pyserial==3.5
PySocks==1.7.1
python-augeas==1.1.0
python-barbicanclient==7.0.0
python-cinderclient==9.6.0
python-congressclient==2.0.1
python-dateutil==2.8.2
python-designateclient==6.1.0
python-dotenv==1.1.1
python-glanceclient==4.7.0
python-heatclient==4.0.0
python-ironic-inspector-client==5.2.0
python-ironicclient==5.8.0
python-keystoneclient==5.5.0
python-linux-procfs==0.7.3
python-magnumclient==4.4.0
python-meh==0.52
python-mistralclient==5.3.0
python-multipart==0.0.20
python-muranoclient==2.8.0
python-neutronclient==11.3.1
python-novaclient==18.6.0
python-octaviaclient==3.8.0
python-openstackclient==7.2.1
python-pam==2.0.2
python-saharaclient==4.2.0
python-searchlightclient==2.1.1
python-senlinclient==3.1.0
python-swiftclient==4.6.0
python-troveclient==8.6.0
python-vitrageclient==5.1.1
python-watcherclient==4.5.0
python-zaqarclient==2.8.0
python-zunclient==5.1.0
pytz==2025.2
pyudev==0.24.3
pyvis==0.3.2
pyxdg==0.27
PyYAML==6.0.1
pyynl @ file:///builddir/build/BUILD/kernel-6.15.7-build/kernel-6.15.7/linux-6.15.7-100.fc41.x86_64/tools/net/ynl
qutebrowser==3.5.0
rapidfuzz==3.5.2
referencing==0.35.1
regex==2023.12.25
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requestsexceptions==1.4.0
resolvelib==1.0.1
rfc3339-validator==0.1.4
rfc3986==2.0.0
rich==13.9.4
rich-rst==1.3.1
rpds-py==0.21.0
rpm==4.20.1
rsa==4.9.1
s3transfer==0.13.1
scour==0.38.2
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
semantic-version==2.10.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
shellingham==1.5.4
shtab==1.7.1
simpleaudio==1.0.4
simplejson==3.19.2
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.9.1
soupsieve==2.7
SQLAlchemy==2.0.42
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.2
stevedore==5.3.0
sympy==1.14.0
systemd-python==235
Tempita==0.5.2
tenacity==8.5.0
termcolor==2.3.0
tiktoken==0.5.2
tinycss2==1.3.0
tldr==3.4.0
tokenizers==0.21.4
tomli==2.2.1
tomli_w==1.2.0
tox==4.25.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.9.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==1.26.20
urwid==2.6.14
userpath==1.9.2
uv==0.8.4
uvicorn==0.35.0
uvloop==0.21.0
virt-lightning==2.3.2
virtualenv==20.29.3
warlock==2.0.1
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
WebOb==1.8.9
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.16.0
wxPython==4.2.1
xkbregistry==0.3
yaql==3.0.0
yarl==1.20.1
zipp==3.23.0
